# openapi.yaml
openapi: 3.0.0
info:
  description: "A RESTful API for managing recipes, ingredients, and users."
  title: Recipe API
  version: 1.0.0
servers:
- url: https://api.example.com/v1
security:
- bearerAuth: []
paths:
  /users:
    get:
      responses:
        "200":
          description: A list of users
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Get a list of users
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "201":
          description: User created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create a new user
  /users/{userId}:
    delete:
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: User deleted
        "401":
          description: Unauthorized
        "404":
          description: User not found
        "500":
          description: Internal Server Error
      summary: Delete a user by ID
    get:
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: A user
        "401":
          description: Unauthorized
        "404":
          description: User not found
        "500":
          description: Internal Server Error
      summary: Get a user by ID
    put:
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: User updated
        "400":
          description: Bad Request
        "404":
          description: User not found
        "500":
          description: Internal Server Error
      summary: Update a user by ID
  /recipes:
    get:
      responses:
        "200":
          description: A list of recipes
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Get a list of recipes
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Recipe"
        required: true
      responses:
        "201":
          description: Recipe created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create a new recipe
  /recipes/{recipeId}:
    delete:
      parameters:
      - explode: false
        in: path
        name: recipeId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Recipe deleted
        "401":
          description: Unauthorized
        "404":
          description: Recipe not found
        "500":
          description: Internal Server Error
      summary: Delete a recipe by ID
    get:
      parameters:
      - explode: false
        in: path
        name: recipeId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: A recipe
        "401":
          description: Unauthorized
        "404":
          description: Recipe not found
        "500":
          description: Internal Server Error
      summary: Get a recipe by ID
    put:
      parameters:
      - explode: false
        in: path
        name: recipeId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Recipe"
        required: true
      responses:
        "200":
          description: Recipe updated
        "400":
          description: Bad Request
        "404":
          description: Recipe not found
        "500":
          description: Internal Server Error
      summary: Update a recipe by ID
  /recipes/{recipeId}/ingredients:
    post:
      parameters:
      - explode: false
        in: path
        name: recipeId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Ingredient"
        required: true
      responses:
        "201":
          description: Ingredient added
        "400":
          description: Bad Request
        "404":
          description: Recipe not found
        "500":
          description: Internal Server Error
      summary: Add an ingredient to a recipe
components:
  schemas:
    User:
      example:
        password: password
        id: id
        email: email
        username: username
      properties:
        id:
          title: id
          type: string
        username:
          title: username
          type: string
        email:
          title: email
          type: string
        password:
          title: password
          type: string
      required:
      - email
      - id
      - password
      - username
      title: User
      type: object
    Recipe:
      example:
        user_id: user_id
        description: description
        ingredients:
        - unit: unit
          quantity: quantity
          name: name
          id: id
        - unit: unit
          quantity: quantity
          name: name
          id: id
        id: id
        title: title
      properties:
        id:
          title: id
          type: string
        title:
          title: title
          type: string
        description:
          title: description
          type: string
        user_id:
          title: user_id
          type: string
        ingredients:
          items:
            $ref: "#/components/schemas/Ingredient"
          title: ingredients
          type: array
      required:
      - description
      - id
      - ingredients
      - title
      - user_id
      title: Recipe
      type: object
    Ingredient:
      example:
        unit: unit
        quantity: quantity
        name: name
        id: id
      properties:
        id:
          title: id
          type: string
        name:
          title: name
          type: string
        quantity:
          title: quantity
          type: string
        unit:
          title: unit
          type: string
      required:
      - id
      - name
      - quantity
      - unit
      title: Ingredient
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http